requirements.txt
PyYAML==6.0.1
slixmpp==1.8.5